+++ src/index.ts
@@ -2,6 +2,8 @@ export * from "./lex.js";
 export * from "./params.js";
 export * from "./meta.js";
 export * from "./document.js";
+export * from "./expr/parse.js";
+export * from "./expr/tables.js";
 export * from "./firmware.js";
 export * from "./commands/g10.js";
 export * from "./commands/toolParams.js";